requirements.txt
fastapi
pydantic
faiss-cpu
numpy
sentence-transformers
rapidfuzz